+++ app/controllers/rates_controller.rb
@@ -1,9 +1,14 @@
 class RatesController < ApplicationController
   def show
-    @rates = Rate.get_rates('USDJPY', 5)
+    pair = params[:pair] || 'USDJPY'
+    interval = params[:interval] || '5-min'
+    @rates = Rate.get_rates(pair, interval)
   end
 
   def update
-    @rates = Rate.get_rates('USDJPY', 5)
+    @rates = Rate.get_rates(params[:pair], params[:interval])
+    respond_to do |format|
+      format.js
+    end
   end
 end