+++ app/models/rate.rb
@@ -3,12 +3,18 @@ class Rate < ActiveRecord::Base
   scope :pair, ->(pair) { where(:pair => pair) }
 
   def self.get_rates(pair, interval)
-    return nil if Rate.pair(pair).select(:pair).distinct.count(:pair) == 0
-    return nil unless interval && interval > 0
+    return nil if Rate.pair(pair).size == 0
+    return nil unless interval
 
     [].tap do |arr|
-      Rate.pair(pair).interval("#{interval}-min").order('to_date DESC').select('to_date, open, close, high, low').limit(30).each do |record|
-        arr << [(record.to_date + 1).strftime('%Y-%m-%d %H:%M:%S'), record.open, record.high, record.low, record.close]
+      Rate.pair(pair).interval(interval).order('to_date DESC').select('to_date, open, close, high, low').limit(100).each do |record|
+        arr << {
+          :time => (record.to_date + 1).strftime('%Y-%m-%d %H:%M:%S'),
+          :open => record.open,
+          :high => record.high,
+          :low => record.low,
+          :close => record.close,
+        }
       end
     end
   end