+++ db/migrate/20160504154106_add_pair_index_to_rates.rb
@@ -1,0 +1,5 @@
+class AddPairIndexToRates < ActiveRecord::Migration
+  def change
+    add_index :rates, :pair
+  end
+end