+++ script/tweets.delete.rb
@@ -1,4 +1,5 @@
 require 'date'
+require 'mysql2'
 
 ENV['TZ'] = 'UTC'
 today = ARGV[0] ? Date.parse(ARGV[0]) : Date.today
@@ -7,6 +8,14 @@ query = <<"EOF"
 DELETE FROM
   tweets
 WHERE
-  DATE(created_at) < '#{(today << 2).strftime('%F')}'
+  created_at < '#{(today << 2).strftime('%Y-%m-%d 00:00:00')}'
 EOF
-`mysql --user=root --password=7QiSlC?4 regulus -e "#{query}"`
+client = Mysql2::Client.new(:host => "localhost", :username => "root", :password => "7QiSlC?4", :database => 'regulus')
+client.query(query)
+client.close
+
+puts [
+  "[#{Time.now.strftime('%Y-%m-%d %H:%M:%S')}]",
+  '[delete]',
+  "{date: #{(today << 2).strftime('%F')}}",
+].join(' ')