+++ spec/rails_helper.rb
@@ -5,6 +5,8 @@ require File.expand_path('../../config/environment', __FILE__)
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'spec_helper'
 require 'rspec/rails'
+require 'capybara/rails'
+require 'capybara/rspec'
 # Add additional requires below this line. Rails is not loaded until this point!
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
@@ -19,12 +21,37 @@ require 'rspec/rails'
 # of increasing the boot-up time by auto-requiring all files in the support
 # directory. Alternatively, in the individual `*_spec.rb` files, manually
 # require only the support files necessary.
-#
-# Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
+
+Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
 
 # Checks for pending migration and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove this line.
 ActiveRecord::Migration.maintain_test_schema!
+
+Capybara.app_host = 'http://160.16.66.112:888'
+Capybara.javascript_driver = :webkit
+Headless.new.start
+
+module Capybara
+  class << self
+    alias_method :old_reset_sessions!, :reset_sessions!
+    def reset_sessions!; end
+  end
+end
+
+Capybara::Webkit.configure do |config|
+  config.block_unknown_urls
+  config.allow_url('http://160.16.66.112:888')
+end
+
+module Capybara::Webkit
+  class Driver < Capybara::Driver::Base
+    def reset!; end
+    def browser
+      @browser
+    end
+  end
+end
 
 RSpec.configure do |config|
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
@@ -54,4 +81,6 @@ RSpec.configure do |config|
   config.filter_rails_from_backtrace!
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
+
+  config.include Capybara::DSL
 end