+++ spec/support/modules/examples/common.rb
@@ -1,0 +1,5 @@
+# coding: utf-8
+
+shared_examples 'ステータスコードが正しいこと' do
+  it { expect(@res.status).to eq(200) }
+end