+++ spec/support/requests/contexts.rb
@@ -1,0 +1,5 @@
+# coding: utf-8
+
+shared_context '認証情報を入力する' do
+  before(:each) { page.driver.browser.authenticate('dev', '.dev') }
+end