+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 regulus
 =======
 
+- 1.3.0
+  - [UPDATE] show rate info on display
+  - [UPDATE] refactor javascript
+
+- 1.2.0
+  - [UPDATE] add select tag for pair & interval
+  - [UPDATE] use d3.js instead of jqplot
+  - [UPDATE] use gem 'mysql2' in external scripts
+
+- 1.1.1
+  - [UPDATE] write style to scss
+  - [UPDATE] change currencies to rates
+  - [UPDATE] modify interval for updating tweets
+
+- 1.1.0
+  - [UPDATE] create each view for rates, tweets and articles
+
 - 1.0.3
   - [UPDATE] scripts
   - [ADD] rspec